third_party/blink/web_tests/fast/dom/navigator-cookieEnabled-no-crash-expected.txt

CONSOLE WARNING: While navigator.cookieEnabled does return true for this 378604901.
If this page appears without a crash, the test has passed.
